OpenWeatherMap_etl/requirements.txt
db-dtypes==1.4.2
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.23.3
pandas==2.2.3
pyarrow==19.0.1
requests==2.32.3 


